+++ core/server/api/canary/utils/serializers/output/utils/post-gating.js
