+++ core/server/api/v2/utils/serializers/output/utils/post-gating.js
